Grader-1/pyproject.toml
[project]
name = "grader-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.1",
    "azure-servicebus>=7.14.2",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "vllm>=0.11.0",
    "websocket>=0.2.1",
    "websocket-client>=1.9.0",
]
